integrated into myscript.py and scripts.db

+++ myscript.py
@@ -21,6 +21,8 @@ class ScriptScribe(object):
             print('{} is not found.'.format(dbFile))
             sys.exit()
         
+        self.run = True
+
         self.script_arguments = ''
         if len(sys.argv) == 1:
             self.script = None
@@ -34,43 +36,78 @@ class ScriptScribe(object):
 
     def confirm_to_overwrite(self, afile):
 
-        if os.path.exists(afile):
-            answer = input('"{}" already exists. Are you sure to overwrite it? [y/N] '.format(afile))
-            if answer.lower() == 'y':
-                os.remove(afile)
-                return True
-            else:
-                return False
-        else:
+        if not self.run:
             return True
 
-
-    def write_from_database(self):
-
-        filename = self.database.get(self.script, 'output', fallback=None)
-        if filename is None:
+        if afile is None:
             print('Ensure the script database file is set properly.')
-            return 
+            return False
+        else:
+            if os.path.exists(afile):
+                answer = input('"{}" already exists. Are you sure to overwrite it? [y/N] '.format(afile))
+                if answer.lower() == 'y':
+                    os.remove(afile)
+                    return True
+                else:
+                    return False
+            else:
+                return True
+
+
+    def write_from_database(self, source, filename):        
 
         if self.confirm_to_overwrite(filename):
             try:
-                content = self.database.get(self.script, 'content')
-                if os.path.splitext(filename)[1] != '.cmd':
-                    content = content.replace('`', '')
-                with open(filename, mode='w', encoding='utf-8') as f:
-                    f.write(content)
+                content = self.database.get(self.script, source)                
             except:
                 print('Ensure the script database file is set properly.')
+                return False 
+            
+            if os.path.splitext(filename)[1] != '.cmd':
+                content = content.replace('`', '')
+            with open(filename, mode='w', encoding='utf-8') as f:
+                f.write(content)
+            return True
+        else:
+            return False   
+
+        
+    def pick_script(self):
+
+        if self.run:
+            if not self.database.has_section(self.script):
+                print('"{}" does not exist.'.format(self.script))
                 return
 
-        cmd = '{} {}'.format(filename, self.script_arguments)
-        if os.path.splitext(filename)[1] == '.ps1':
-            cmd = 'powershell.exe ./' + cmd        
-        # print(cmd)
-        os.system(cmd)        
+        options=self.database.options(self.script)
+        for i, option in enumerate(options):
+            if 'output' in option:
+                filename = self.database.get(self.script, option)
+                source = option.split('_')[0]
+                if not self.write_from_database(source, filename):
+                    return
+
+        if self.run:
+            cmd = '{} {}'.format(filename, self.script_arguments)
+            if os.path.splitext(filename)[1] == '.ps1':
+                cmd = 'powershell.exe ./' + cmd        
+            # print(cmd)
+            os.system(cmd)
+
+
+    def burst_scripts(self):
+
+        self.run = False
+        scripts = sorted(self.database.sections(), key=str.casefold)
+
+        for i in scripts:
+            self.script = i
+            self.pick_script()
 
 
     def enumerate_scripts(self):
+
+        print('Specify "ALL" to  take out all scripts.')
 
         scripts = sorted(self.database.sections(), key=str.casefold)
 
@@ -85,8 +122,10 @@ if __name__ == '__main__':
     SS = ScriptScribe()
     if SS.script is None:
         SS.enumerate_scripts()
+    elif SS.script == 'ALL':
+        SS.burst_scripts()
     else:
-        SS.write_from_database()
+        SS.pick_script()
